refactor: packageing 리팩토링 및 가독성 개선
