+++ tomcat/src/main/java/nextstep/jwp/service/LoginService.java
@@ -1,5 +1,6 @@
-package nextstep.jwp;
+package nextstep.jwp.service;
 
+import nextstep.jwp.LoginFailureException;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 import org.apache.coyote.http11.Http11Processor;
